requirements.txt
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-http>=1.20.0
mcp>=0.9.0
httpx>=0.25.0
pydantic>=2.0.0
google-genai>=0.3.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
